requirements.txt
anthropic>=0.50
python-dotenv
geopandas
rasterio
shapely
pandas
numpy
requests
matplotlib
scipy
pyproj
